feature: conditionally show the annual report at /2020
