+++ src/pages/2020/comingSoon.js
@@ -1,0 +1,13 @@
+/** @jsx jsx */
+
+import { jsx, Flex, Text } from 'theme-ui'
+
+const ComingSoon = () => {
+	return (
+		<Flex>
+			<Text>Coming Soon</Text>
+		</Flex>
+	)
+}
+
+export default ComingSoon
